+++ src/main/java/com/lazerycode/selenium/download/DownloadHandler.java
@@ -1,8 +1,9 @@
 package com.lazerycode.selenium.download;
 
-import com.lazerycode.selenium.repository.BinaryType;
-import com.lazerycode.selenium.extract.ExtractFilesFromArchive;
-import com.lazerycode.selenium.repository.FileDetails;
+import com.lazerycode.selenium.hash.HashType;
+import com.lazerycode.selenium.repository.DriverContext;
+import com.lazerycode.selenium.repository.DriverDetails;
+import com.lazerycode.selenium.repository.DriverMap;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.log4j.Logger;
 import org.apache.maven.plugin.MojoExecutionException;
@@ -11,12 +12,15 @@ import org.apache.maven.plugin.MojoFailureException;
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.net.URL;
 import java.util.Map;
+
+import static com.lazerycode.selenium.extract.ExtractFilesFromArchive.extractFileFromArchive;
 
 public class DownloadHandler {
 
     private static final Logger LOG = Logger.getLogger(DownloadHandler.class);
-    private final Map<String, FileDetails> filesToDownload;
+    private final DriverMap filesToDownload;
     private final File rootStandaloneServerDirectory;
     private final File downloadedZipFileDirectory;
     protected final int fileDownloadRetryAttempts;
@@ -24,7 +28,7 @@ public class DownloadHandler {
     private boolean checkFileHash = true;
     private FileDownloader fileDownloader;
 
-    public DownloadHandler(File rootStandaloneServerDirectory, File downloadedZipFileDirectory, int fileDownloadRetryAttempts, int fileDownloadConnectTimeout, int fileDownloadReadTimeout, Map<String, FileDetails> filesToDownload, boolean overwriteFilesThatExist, boolean checkFileHash, boolean useSystemProxy) throws MojoFailureException {
+    public DownloadHandler(File rootStandaloneServerDirectory, File downloadedZipFileDirectory, int fileDownloadRetryAttempts, int fileDownloadConnectTimeout, int fileDownloadReadTimeout, DriverMap filesToDownload, boolean overwriteFilesThatExist, boolean checkFileHash, boolean useSystemProxy) throws MojoFailureException {
         this.rootStandaloneServerDirectory = rootStandaloneServerDirectory;
         this.downloadedZipFileDirectory = downloadedZipFileDirectory;
         this.filesToDownload = filesToDownload;
@@ -37,40 +41,11 @@ public class DownloadHandler {
         fileDownloader.setConnectTimeout(fileDownloadConnectTimeout);
     }
 
-    public void ensureStandaloneExecutableFilesExist() throws MojoFailureException, MojoExecutionException, IOException, URISyntaxException {
-        LOG.info("Archives will be downloaded to '" + this.downloadedZipFileDirectory.getAbsolutePath() + "'");
-        LOG.info("Standalone executable files will be extracted to '" + this.rootStandaloneServerDirectory + "'");
-        LOG.info(" ");
-        LOG.info("Preparing to download Selenium Standalone Executable Binaries...");
-        for (Map.Entry<String, FileDetails> fileToDownload : this.filesToDownload.entrySet()) {
-            LOG.info(" ");
-            String currentFileAbsolutePath = this.downloadedZipFileDirectory + File.separator + FilenameUtils.getName(fileToDownload.getValue().getFileLocation().getFile());
-            File desiredFile = new File(currentFileAbsolutePath);
-            File fileToUnzip = null;
-            LOG.info("Archive file '" + desiredFile.getName() + "' exists   : " + desiredFile.exists());
-            if (desiredFile.exists()) {
-                if (checkFileHash) {
-                    FileHashChecker fileHashChecker = new FileHashChecker(desiredFile);
-                    fileHashChecker.setExpectedHash(fileToDownload.getValue().getHash(), fileToDownload.getValue().getHashType());
-                    boolean fileIsValid = fileHashChecker.fileIsValid();
-                    LOG.info("Archive file '" + desiredFile.getName() + "' is valid : " + fileIsValid);
-                    if (fileIsValid) {
-                        fileToUnzip = desiredFile;
-                    }
-                } else {
-                    fileToUnzip = desiredFile;
-                }
-            }
-            if (fileToUnzip == null) {
-                fileToUnzip = downloadValidFile(fileToDownload.getValue());
-            }
-            String extractedFileLocation = this.rootStandaloneServerDirectory.getAbsolutePath() + File.separator + fileToDownload.getKey();
-            String binaryForOperatingSystem = fileToDownload.getKey().replace("\\", "/").split("/")[1].toUpperCase();  //TODO should really store the OSType we have extracted somewhere rather than doing this hack!
-            LOG.debug("Detected a binary for OSType: " + binaryForOperatingSystem);
-            if (ExtractFilesFromArchive.extractFileFromArchive(fileToUnzip, extractedFileLocation, this.overwriteFilesThatExist, BinaryType.valueOf(binaryForOperatingSystem))) {
-                LOG.info("File(s) copied to " + extractedFileLocation);
-            }
-        }
+    private boolean checkFileHash(File fileToCheck, String hash, HashType hashType) throws IOException, MojoExecutionException {
+        FileHashChecker fileHashChecker = new FileHashChecker(fileToCheck);
+        fileHashChecker.setExpectedHash(hash, hashType);
+
+        return fileHashChecker.fileIsValid();
     }
 
     /**
@@ -79,33 +54,65 @@ public class DownloadHandler {
      * @return File
      * @throws MojoExecutionException
      */
-    File downloadValidFile(FileDetails fileDetails) throws MojoExecutionException, IOException, URISyntaxException {
-        final String filename = FilenameUtils.getName(fileDetails.getFileLocation().getFile());
+    protected File downloadFile(DriverDetails driverDetails, boolean shouldWeCheckFileHash) throws MojoExecutionException, IOException, URISyntaxException {
+
+        URL remoteFileLocation = driverDetails.fileLocation;
+
+        final String filename = FilenameUtils.getName(remoteFileLocation.getFile());
         for (int retryAttempts = 1; retryAttempts <= this.fileDownloadRetryAttempts; retryAttempts++) {
-
-            File downloadedFile = fileDownloader.attemptToDownload(fileDetails.getFileLocation());
-
+            File downloadedFile = fileDownloader.attemptToDownload(remoteFileLocation);
             if (null != downloadedFile) {
-                if (!checkFileHash) {
+                if (!shouldWeCheckFileHash || (shouldWeCheckFileHash && checkFileHash(downloadedFile, driverDetails.hash, driverDetails.hashType))) {
+                    LOG.info("Archive file '" + downloadedFile.getName() + "' is valid : true");
                     return downloadedFile;
-                }
-
-                FileHashChecker fileHashChecker = new FileHashChecker(downloadedFile);
-                fileHashChecker.setExpectedHash(fileDetails.getHash(), fileDetails.getHashType());
-                boolean isFileValid = fileHashChecker.fileIsValid();
-                LOG.info("Archive file '" + downloadedFile.getName() + "' is valid : " + isFileValid);
-                if (isFileValid) {
-                    return downloadedFile;
+                } else {
+                    LOG.info("Archive file '" + downloadedFile.getName() + "' is valid : false");
                 }
             }
-
             LOG.info("Problem downloading '" + filename + "'... ");
-
             if (retryAttempts < this.fileDownloadRetryAttempts) {
                 LOG.info("Retry attempt " + (retryAttempts) + " for '" + filename + "'");
             }
         }
 
         throw new MojoExecutionException("Unable to successfully download '" + filename + "'!");
+    }
+
+    public void ensureStandaloneExecutableFilesExist() throws MojoFailureException, MojoExecutionException, IOException, URISyntaxException {
+        LOG.info("Archives will be downloaded to '" + this.downloadedZipFileDirectory.getAbsolutePath() + "'");
+        LOG.info("Standalone executable files will be extracted to '" + this.rootStandaloneServerDirectory + "'");
+        LOG.info(" ");
+        LOG.info("Preparing to download Selenium Standalone Executable Binaries...");
+
+        for (final DriverContext driverContext : filesToDownload.getKeys()) {
+            Map<String, DriverDetails> driverDetails = filesToDownload.getMapForDriverContext(driverContext);
+            for (String version : driverDetails.keySet()) {
+                URL fileLocation = driverDetails.get(version).fileLocation;
+                String hash = driverDetails.get(version).hash;
+                HashType hashType = driverDetails.get(version).hashType;
+                String localZipFileAbsolutePath = this.downloadedZipFileDirectory + File.separator + FilenameUtils.getName(fileLocation.getFile());
+                File localZipFile = new File(localZipFileAbsolutePath);
+                boolean fileNeedsToBeDownloaded = true;
+
+                if (localZipFile.exists()) {
+                    if (checkFileHash) {
+                        if (checkFileHash(localZipFile, hash, hashType)) {
+                            fileNeedsToBeDownloaded = false;
+                        } else {
+                            fileNeedsToBeDownloaded = false;
+                        }
+                    }
+                }
+
+                if (fileNeedsToBeDownloaded) {
+                    localZipFile = downloadFile(driverDetails.get(version), checkFileHash);
+                }
+
+                String extractedFileLocation = this.rootStandaloneServerDirectory.getAbsolutePath() + File.separator + driverContext.buildExtractionPathFromDriverContext();
+                if (extractFileFromArchive(localZipFile, extractedFileLocation, this.overwriteFilesThatExist, driverContext.getBinaryTypeForContext())) {
+                    LOG.info("File(s) copied to " + extractedFileLocation);
+                }
+            }
+        }
     }
 }