+++ src/test/java/com/lazerycode/selenium/download/DownloadHandlerTest.java
@@ -1,6 +1,7 @@
 package com.lazerycode.selenium.download;
 
-import com.lazerycode.selenium.repository.FileDetails;
+import com.lazerycode.selenium.repository.DriverDetails;
+import com.lazerycode.selenium.repository.DriverMap;
 import org.apache.commons.io.FileUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -9,9 +10,13 @@ import org.junit.*;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
-import java.util.HashMap;
 
 import static com.lazerycode.selenium.hash.HashType.SHA1;
+import static com.lazerycode.selenium.repository.BinaryType.GOOGLECHROME;
+import static com.lazerycode.selenium.repository.BinaryType.PHANTOMJS;
+import static com.lazerycode.selenium.repository.DriverContext.binaryDataFor;
+import static com.lazerycode.selenium.repository.OperatingSystem.OSX;
+import static com.lazerycode.selenium.repository.SystemArchitecture.ARCHITECTURE_64_BIT;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import static org.hamcrest.core.IsNot.not;
@@ -27,14 +32,17 @@ public class DownloadHandlerTest {
     private final int oneRetryAttempt = 1;
     private final int connectTimeout = 15000;
     private final int readTimeout = 15000;
+    private final boolean overwriteFilesThatExist = true;
     private final boolean doNotOverwriteFilesThatExist = false;
     private final boolean doNotCheckFileHashes = false;
     private final boolean checkFileHashes = true;
+    private final boolean doNotUseSystemProxy = false;
 
     private static final String validSHA1Hash = "8604c05969a0eefa0edf0d71ae809310832afdc7";
     private static final JettyServer localWebServer = new JettyServer();
     private static final String webServerURL = webServerAddress + ":" + webServerPort;
-    private static FileDetails validFileDetails;
+    private static DriverMap validDriverMap = new DriverMap();
+    private static DriverDetails validFileDetails = new DriverDetails();
     private static File downloadDirectory;
     private static File rootStandaloneServerDirectory;
     private static URL downloadZipURL;
@@ -43,7 +51,10 @@ public class DownloadHandlerTest {
     public static void start() throws Exception {
         localWebServer.startJettyServer(webServerPort);
         downloadZipURL = new URL(webServerURL + "/files/download.zip");
-        validFileDetails = new FileDetails(downloadZipURL, SHA1, validSHA1Hash);
+        validFileDetails.fileLocation = downloadZipURL;
+        validFileDetails.hashType = SHA1;
+        validFileDetails.hash = validSHA1Hash;
+        validDriverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT)).put("2.13", validFileDetails);
     }
 
     @AfterClass
@@ -65,47 +76,67 @@ public class DownloadHandlerTest {
 
     @Test(expected = MojoExecutionException.class)
     public void downloadAFile() throws Exception {
-        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, oneRetryAttempt, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, false);
-        downloadTestFile.downloadValidFile(new FileDetails(downloadZipURL, SHA1, ""));
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = downloadZipURL;
+        fileDetails.hashType = SHA1;
+        fileDetails.hash = "";
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, oneRetryAttempt, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
+        downloadTestFile.downloadFile(fileDetails, checkFileHashes);
     }
 
     @Test(expected = MojoExecutionException.class)
     public void invalidHashCheck() throws Exception {
-        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, oneRetryAttempt, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, false);
-        downloadTestFile.downloadValidFile(new FileDetails(downloadZipURL, SHA1, "invalidHash"));
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = downloadZipURL;
+        fileDetails.hashType = SHA1;
+        fileDetails.hash = "invalidHash";
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, oneRetryAttempt, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
+        downloadTestFile.downloadFile(fileDetails, checkFileHashes);
     }
 
     @Test
     public void hashCheck() throws Exception {
-        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, oneRetryAttempt, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, false);
-        downloadTestFile.downloadValidFile(validFileDetails);
+        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, oneRetryAttempt, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
+        downloadTestFile.downloadFile(validFileDetails, checkFileHashes);
 
         assertThat(new File(expectedDownloadedFilePath).exists(), is(equalTo(true)));
     }
 
     @Test(expected = MojoExecutionException.class)
     public void tryToDownloadAnInvalidFile() throws Exception {
-        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, 3, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, false);
-        downloadTestFile.downloadValidFile(new FileDetails(new URL(webServerURL + "/files/null/download.zip"), SHA1, validSHA1Hash));
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = new URL(webServerURL + "/files/null/download.zip");
+        fileDetails.hashType = SHA1;
+        fileDetails.hash = validSHA1Hash;
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        DownloadHandler downloadTestFile = new DownloadHandler(null, downloadDirectory, 3, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
+        downloadTestFile.downloadFile(fileDetails, checkFileHashes);
     }
 
     @Test(expected = MojoFailureException.class)
     public void specifyAFileInsteadOfADirectory() throws Exception {
-        DownloadHandler downloadTestFile = new DownloadHandler(null, File.createTempFile("foo", "bar"), 3, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, false);
-        downloadTestFile.downloadValidFile(new FileDetails(new URL(webServerURL + "/files/null/download.zip"), SHA1, validSHA1Hash));
+        DownloadHandler downloadTestFile = new DownloadHandler(null, File.createTempFile("foo", "bar"), 3, connectTimeout, readTimeout, null, doNotOverwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
+        downloadTestFile.downloadFile(validFileDetails, checkFileHashes);
     }
 
     @Test
     public void filesAreExtractedIntoTheCorrectStandaloneServerPathAndCanBeOverwritten() throws Exception {
-        String downloadPath = "os/phantomjs/32bit/1";
-        File expectedDownloadedFile = new File(rootStandaloneServerDirectoryPath + File.separator + downloadPath + File.separator + "phantomjs");
-        FileDetails testFileDetails = new FileDetails(downloadZipURL, SHA1, validSHA1Hash);
-        HashMap<String, FileDetails> fileDownloadList = new HashMap<String, FileDetails>();
-        fileDownloadList.put(downloadPath, testFileDetails);
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = downloadZipURL;
+        fileDetails.hashType = SHA1;
+        fileDetails.hash = validSHA1Hash;
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, PHANTOMJS, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        File expectedDownloadedFile = new File(rootStandaloneServerDirectoryPath + File.separator + "mac" + File.separator + "phantomjs" + File.separator + "64bit" + File.separator + "phantomjs");
 
         assertThat(expectedDownloadedFile.exists(), is(equalTo(false)));
 
-        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, fileDownloadList, true, checkFileHashes, false);
+        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, driverMap, overwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
         downloadTestFile.ensureStandaloneExecutableFilesExist();
 
         assertThat(expectedDownloadedFile.exists(), is(equalTo(true)));
@@ -120,15 +151,17 @@ public class DownloadHandlerTest {
 
     @Test
     public void fileCanBeDownloadedIfThereIsNoHashInFileDownloadListFileDetails() throws Exception {
-        String downloadPath = "os/phantomjs/32bit/1";
-        File expectedDownloadedFile = new File(rootStandaloneServerDirectoryPath + File.separator + downloadPath + File.separator + "phantomjs");
-        FileDetails testFileDetails = new FileDetails(downloadZipURL, null, null);
-        HashMap<String, FileDetails> fileDownloadList = new HashMap<String, FileDetails>();
-        fileDownloadList.put(downloadPath, testFileDetails);
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = downloadZipURL;
+        fileDetails.hashType = null;
+        fileDetails.hash = null;
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, PHANTOMJS, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        File expectedDownloadedFile = new File(rootStandaloneServerDirectoryPath + File.separator + "mac" + File.separator + "phantomjs" + File.separator + "64bit" + File.separator + "phantomjs");
 
         assertThat(expectedDownloadedFile.exists(), is(equalTo(false)));
 
-        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, fileDownloadList, true, doNotCheckFileHashes, false);
+        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, driverMap, overwriteFilesThatExist, doNotCheckFileHashes, doNotUseSystemProxy);
         downloadTestFile.ensureStandaloneExecutableFilesExist();
 
         assertThat(expectedDownloadedFile.exists(), is(equalTo(true)));
@@ -136,33 +169,25 @@ public class DownloadHandlerTest {
 
     @Test(expected = MojoExecutionException.class)
     public void errorThrownIfThereIsNoHashInFileDownloadListFileDetailsAndHashShouldBeChecked() throws Exception {
-        String downloadPath = "os/phantomjs/32bit/1";
-        File expectedDownloadedFile = new File(rootStandaloneServerDirectoryPath + File.separator + downloadPath + File.separator + "phantomjs");
-        FileDetails testFileDetails = new FileDetails(downloadZipURL, SHA1, null);
-        HashMap<String, FileDetails> fileDownloadList = new HashMap<String, FileDetails>();
-        fileDownloadList.put(downloadPath, testFileDetails);
-
-        assertThat(expectedDownloadedFile.exists(), is(equalTo(false)));
-
-        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, fileDownloadList, true, checkFileHashes, false);
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = downloadZipURL;
+        fileDetails.hashType = SHA1;
+        fileDetails.hash = null;
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, driverMap, overwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
         downloadTestFile.ensureStandaloneExecutableFilesExist();
-
-        assertThat(expectedDownloadedFile.exists(), is(equalTo(true)));
     }
 
     @Test(expected = NullPointerException.class)
     public void errorThrownIfThereIsNoHashTypeInFileDownloadListFileDetailsAndHashShouldBeChecked() throws Exception {
-        String downloadPath = "os/phantomjs/32bit/1";
-        File expectedDownloadedFile = new File(rootStandaloneServerDirectoryPath + File.separator + downloadPath + File.separator + "phantomjs");
-        FileDetails testFileDetails = new FileDetails(downloadZipURL, null, "foo");
-        HashMap<String, FileDetails> fileDownloadList = new HashMap<String, FileDetails>();
-        fileDownloadList.put(downloadPath, testFileDetails);
-
-        assertThat(expectedDownloadedFile.exists(), is(equalTo(false)));
-
-        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, fileDownloadList, true, checkFileHashes, false);
+        DriverDetails fileDetails = new DriverDetails();
+        fileDetails.fileLocation = downloadZipURL;
+        fileDetails.hashType = null;
+        fileDetails.hash = validSHA1Hash;
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT)).put("2.13", fileDetails);
+        DownloadHandler downloadTestFile = new DownloadHandler(new File(rootStandaloneServerDirectoryPath), new File(downloadDirectoryPath), oneRetryAttempt, connectTimeout, readTimeout, driverMap, overwriteFilesThatExist, checkFileHashes, doNotUseSystemProxy);
         downloadTestFile.ensureStandaloneExecutableFilesExist();
-
-        assertThat(expectedDownloadedFile.exists(), is(equalTo(true)));
     }
 }