+++ src/test/java/com/lazerycode/selenium/repository/DriverContextTest.java
@@ -2,7 +2,10 @@ package com.lazerycode.selenium.repository;
 
 import org.junit.Test;
 
+import java.io.File;
+
 import static com.lazerycode.selenium.repository.BinaryType.GOOGLECHROME;
+import static com.lazerycode.selenium.repository.DriverContext.binaryDataFor;
 import static com.lazerycode.selenium.repository.OperatingSystem.OSX;
 import static com.lazerycode.selenium.repository.SystemArchitecture.ARCHITECTURE_64_BIT;
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -13,7 +16,24 @@ public class DriverContextTest {
 
     @Test
     public void driverContextWithStringsMapsTpDriverContextWithEnums() {
-        assertThat(DriverContext.binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT),
-                is(equalTo(DriverContext.binaryDataFor("googlechrome", "osx", ARCHITECTURE_64_BIT))));
+        assertThat(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT),
+                is(equalTo(binaryDataFor("osx", "googlechrome", ARCHITECTURE_64_BIT))));
+    }
+
+    @Test
+    public void driverContextCreatesAStringThatCanBeUsedAsAFilePath() {
+        String expectedFilepath = "mac" + File.separator + "googlechrome" + File.separator + "64bit" + File.separator;
+        String filePath = binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT).buildExtractionPathFromDriverContext();
+
+        assertThat(filePath,
+                is(equalTo(expectedFilepath)));
+    }
+
+    @Test
+    public void returnsCorrectBinaryTypeForContext() {
+        BinaryType binaryType = BinaryType.INTERNETEXPLORER;
+
+        assertThat(binaryDataFor(OSX, binaryType, ARCHITECTURE_64_BIT).getBinaryTypeForContext(),
+                is(equalTo(binaryType)));
     }
 }