+++ src/test/java/com/lazerycode/selenium/repository/DriverMapTest.java
@@ -7,9 +7,11 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Map;
 import java.util.NoSuchElementException;
+import java.util.Set;
 import java.util.TreeMap;
 
 import static com.lazerycode.selenium.repository.BinaryType.GOOGLECHROME;
+import static com.lazerycode.selenium.repository.BinaryType.OPERACHROMIUM;
 import static com.lazerycode.selenium.repository.DriverContext.binaryDataFor;
 import static com.lazerycode.selenium.repository.OperatingSystem.OSX;
 import static com.lazerycode.selenium.repository.SystemArchitecture.ARCHITECTURE_64_BIT;
@@ -23,7 +25,7 @@ public class DriverMapTest {
     @Test
     public void willReturnEmptyMap() {
         DriverMap driverMap = new DriverMap();
-        Map versionMap = driverMap.getMapForDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        Map versionMap = driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
 
         assertThat(versionMap,
                 is(instanceOf(TreeMap.class)));
@@ -34,10 +36,10 @@ public class DriverMapTest {
     @Test
     public void willReturnExistingMap() {
         DriverMap driverMap = new DriverMap();
-        Map originalVersionMap = driverMap.getMapForDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        Map originalVersionMap = driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
         originalVersionMap.put("2.13", new DriverDetails());
 
-        Map versionMap = driverMap.getMapForDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        Map versionMap = driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
 
         assertThat(versionMap,
                 is(instanceOf(TreeMap.class)));
@@ -48,30 +50,30 @@ public class DriverMapTest {
     @Test(expected = IllegalArgumentException.class)
     public void willThrowAnExceptionWhenTryingToGetSpecificVersionMapForInvalidContext() {
         DriverMap driverMap = new DriverMap();
-        driverMap.getDetailsForVersionOfDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT), "4");
+        driverMap.getDetailsForVersionOfDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT), "4");
     }
 
     @Test(expected = NoSuchElementException.class)
     public void willThrowAnExceptionWhenTryingToGetAnInvalidVersion() {
         DriverMap driverMap = new DriverMap();
-        Map originalVersionMap = driverMap.getMapForDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        Map originalVersionMap = driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
         originalVersionMap.put("2.13", new DriverDetails());
 
-        driverMap.getDetailsForVersionOfDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT), "4");
+        driverMap.getDetailsForVersionOfDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT), "4");
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void willThrowAnExceptionWhenTryingToGetLatestVersionMapForInvalidContext() {
         DriverMap driverMap = new DriverMap();
-        driverMap.getDetailsForLatestVersionOfDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        driverMap.getDetailsForLatestVersionOfDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
     }
 
     @Test(expected = NoSuchElementException.class)
     public void willThrowAnExceptionWhenTryingToGetAnInvalidLatestVersion() {
         DriverMap driverMap = new DriverMap();
-        driverMap.getMapForDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
 
-        driverMap.getDetailsForLatestVersionOfDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        driverMap.getDetailsForLatestVersionOfDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
     }
 
     @Test
@@ -85,11 +87,11 @@ public class DriverMapTest {
         latestVersion.hash = "latest";
         latestVersion.hashType = HashType.MD5;
         latestVersion.fileLocation = new URL("http://www.example.com/bar/foo");
-        Map originalVersionMap = driverMap.getMapForDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        Map originalVersionMap = driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
         originalVersionMap.put("2.13", oldestVersion);
         originalVersionMap.put("4", latestVersion);
 
-        DriverDetails returnedDetails = driverMap.getDetailsForLatestVersionOfDriverContext(binaryDataFor(GOOGLECHROME, OSX, ARCHITECTURE_64_BIT));
+        DriverDetails returnedDetails = driverMap.getDetailsForLatestVersionOfDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
 
         assertThat(returnedDetails.hash,
                 is(equalTo("latest")));
@@ -97,5 +99,28 @@ public class DriverMapTest {
                 is(equalTo(HashType.MD5)));
         assertThat(returnedDetails.fileLocation,
                 is(equalTo(new URL("http://www.example.com/bar/foo"))));
+    }
+
+    @Test
+    public void willReturnAValidKeySetWithASingleKey() throws MalformedURLException {
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
+
+        Set<DriverContext> allDriverContexts = driverMap.getKeys();
+
+        assertThat(allDriverContexts.size(),
+                is(equalTo(1)));
+    }
+
+    @Test
+    public void willReturnAValidKeySetWithMultipleKeys() throws MalformedURLException {
+        DriverMap driverMap = new DriverMap();
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, GOOGLECHROME, ARCHITECTURE_64_BIT));
+        driverMap.getMapForDriverContext(binaryDataFor(OSX, OPERACHROMIUM, ARCHITECTURE_64_BIT));
+
+        Set<DriverContext> allDriverContexts = driverMap.getKeys();
+
+        assertThat(allDriverContexts.size(),
+                is(equalTo(2)));
     }
 }